Add files via upload
This project will analyze and evaluate the success, investment models, and trends of the TV show Shark Tank through 495 companies appearing in 122 episodes of the first six seasons of the show. 

By exploring the data collected from the program, the project will answer the question: Which companies are the best and what are the trends that lead to successful deals in Shark Tank?